+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v0.2.0 (in development)
 -----------------------
 - Increased MSRV to 1.88
+- Added a `ClientBuilder` for setting a client's access token, user agent, base
+  API URL, API version header value, and/or "Accept" header value
 
 v0.1.1 (2025-06-27)
 -------------------